Add CI workflow for Doxygen documentation
A new workflow file doxygen.yml has been added under the GitHub Actions directory. This workflow triggers on push or pull request to the master branch and runs on the latest Ubuntu runner. It uses the 'doxygenize' action to generate documentation from code using Doxygen and publishes it to a 'gh-pages' branch.

Signed-off-by: Tategoto Azarasi <2724167997@qq.com>