Add armory import/export and update item checks
Extended the ArmyArmory class to support importing and exporting the armory to and from a JSON file. Additionally, increased the robustness of armory item checks to handle cases where the item type is invalid, or the item's StringId is null or empty. Finally, updated the game menu options to include these functionalities.

Signed-off-by: Tategoto Azarasi <2724167997@qq.com>